// src/types/member.ts
export interface MemberForm {
    memberId: string;
    password: string;
    name: string;
    email: string;
}

export interface MemberRequest {
    memberId: string;
    password: string;
    name: string;
    email: string;
}
export interface MemberResponse {
    memberId: string;
    password: string;
    name: string;
    email: string;
}